refactor: util 엔티티 매핑
