+++ core/src/main/java/com/picus/core/domain/studio/entity/UsedCategory.java
@@ -1,9 +1,8 @@
 package com.picus.core.domain.studio.entity;
 
 import com.picus.core.global.common.BaseEntity;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.IdClass;
+import com.picus.core.global.common.category.entity.Category;
+import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -15,13 +14,17 @@ import lombok.NoArgsConstructor;
 public class UsedCategory extends BaseEntity {
 
     @Id
-    private Long studioNo;
+    @ManyToOne
+    @JoinColumn(name = "studio_no", nullable = false)
+    private Studio studio;
 
     @Id
-    private Long categoryId;
+    @ManyToOne
+    @JoinColumn(name = "category_id", nullable = false)
+    private Category category;
 
-    public UsedCategory(Long studioNo, Long categoryId) {
-        this.studioNo = studioNo;
-        this.categoryId = categoryId;
+    public UsedCategory(Studio studio, Category category) {
+        this.studio = studio;
+        this.category = category;
     }
 }
